improved this project (more clean, added BigDecimal)

+++ src/main/java/mate/academy/springboot/datajpa/dto/response/ProductResponseDto.java
@@ -1,5 +1,6 @@
 package mate.academy.springboot.datajpa.dto.response;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -8,7 +9,6 @@ import lombok.Setter;
 public class ProductResponseDto {
     private Long id;
     private String title;
-    private float price;
+    private BigDecimal price;
     private Long categoryId;
-
 }
