+++ src/main/java/mate/academy/springboot/datajpa/service/ProductService.java
@@ -1,5 +1,6 @@
 package mate.academy.springboot.datajpa.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import mate.academy.springboot.datajpa.model.Product;
@@ -13,7 +14,7 @@ public interface ProductService {
 
     void delete(Long id);
 
-    List<Product> getByPriceBetween(double from, double to);
+    List<Product> getByPriceBetween(BigDecimal from, BigDecimal to);
 
     List<Product> getAllByCategories(Map<String, String> params);
 }
